gcp/main.py
from google.cloud import storage
import tensorflow as tf
from PIL import Image
import numpy as np
import logging
model = None
interpreter = None
input_index = None
output_index = None
import pickle
class_names = ["benign","malignant"]

BUCKET_NAME = "malignantvsbenignmodel" # Here you need to put the name of your GCP bucket


def download_blob(bucket_name, source_blob_name, destination_file_name):
    """Downloads a blob from the bucket."""
    storage_client = storage.Client()
    bucket = storage_client.get_bucket(bucket_name)
    blob = bucket.blob(source_blob_name)

    blob.download_to_filename(destination_file_name)

    print(f"Blob {source_blob_name} downloaded to {destination_file_name}.")


def predict(request):
    global model
    if model is None:
        download_blob(
            BUCKET_NAME,
            "models/perfect_model_updated_pickel.pkl",
            "/tmp/perfect_model_updated_pickel.pkl",
        )
        model = pickle.load(open('/tmp/perfect_model_updated_pickel.pkl', 'rb'))

    image = request.files["file"]

    image = np.array(
        Image.open(image).convert("RGB").resize((256, 256)) # image resizing
    )

    image = image/255 # normalize the image in 0 to 1 range

    img_array = tf.expand_dims(image, 0)
    predictions = model.predict(img_array)

    print("Predictions:", predictions)

    predicted_class = class_names[np.argmax(predictions[0])]
    confidence = round(100 * (np.max(predictions[0])), 2)

    return {"class": predicted_class, "confidence": confidence}

