gcp/requirements.txt
tensorflow==2.13.0
google-cloud-storage==1.16.1
Pillow==6.0.0